Tree IR (promote from basic2wasm crate)

+++ src/vm/array.rs
@@ -7,7 +7,7 @@ pub trait Subscript {
         Self: Sized;
 }
 
-impl Subscript for u8 {
+impl Subscript for usize {
     fn to_usize(self, arr: &Array<Self>) -> Option<usize> {
         if self < arr.bound {
             Some(self as usize)
@@ -17,7 +17,7 @@ impl Subscript for u8 {
     }
 }
 
-impl Subscript for [u8; 2] {
+impl Subscript for [usize; 2] {
     fn to_usize(self, arr: &Array<Self>) -> Option<usize> {
         let [m, n] = arr.bound;
         let [i, j] = self;
@@ -93,7 +93,6 @@ impl<I: Subscript> Array<I> {
         }
     }
 
-    #[inline]
     fn grow_to(&mut self, index: usize) {
         while self.values.len() <= index + 1 {
             self.values.push(0.0);
